tests/wpt/tests/html/canvas/element/path-objects/2d.path.ellipse.basics.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.ellipse.basics</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.ellipse.basics</h1>
<p class="desc">Verify canvas throws error when drawing ellipse with negative radii.</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Verify canvas throws error when drawing ellipse with negative radii.");
_addTest(function(canvas, ctx) {

  ctx.ellipse(10, 10, 10, 5, 0, 0, 1, false);
  ctx.ellipse(10, 10, 10, 0, 0, 0, 1, false);
  ctx.ellipse(10, 10, -0, 5, 0, 0, 1, false);
  assert_throws_dom("INDEX_SIZE_ERR", function() { ctx.ellipse(10, 10, -2, 5, 0, 0, 1, false); });
  assert_throws_dom("INDEX_SIZE_ERR", function() { ctx.ellipse(10, 10, 0, -1.5, 0, 0, 1, false); });
  assert_throws_dom("INDEX_SIZE_ERR", function() { ctx.ellipse(10, 10, -2, -5, 0, 0, 1, false); });
  ctx.ellipse(80, 0, 10, 4294967277, Math.PI / -84, -Math.PI / 2147483436, false);

});
</script>
